requirements.txt
requests
sentence-transformers
google-cloud-texttospeech
weaviate-client
